+++ R/provider-databricks.R
@@ -23,18 +23,17 @@
 #' ## Known limitations
 #'
 #' Databricks models do not support images, but they do support structured
-#' outputs. Tool calling support is also very limited at present and is
-#' currently not supported by ellmer.
+#' outputs and tool calls for most models.
 #'
 #' @family chatbots
 #' @param workspace The URL of a Databricks workspace, e.g.
 #'   `"https://example.cloud.databricks.com"`. Will use the value of the
 #'   environment variable `DATABRICKS_HOST`, if set.
-#' @param model `r param_model("databricks-dbrx-instruct")`
+#' @param model `r param_model("databricks-claude-3-7-sonnet")`
 #'
 #'   Available foundational models include:
 #'
-#'   - `databricks-dbrx-instruct` (the default)
+#'   - `databricks-claude-3-7-sonnet` (the default)
 #'   - `databricks-mixtral-8x7b-instruct`
 #'   - `databricks-meta-llama-3-1-70b-instruct`
 #'   - `databricks-meta-llama-3-1-405b-instruct`
@@ -58,7 +57,7 @@ chat_databricks <- function(
 ) {
   check_string(workspace, allow_empty = FALSE)
   check_string(token, allow_empty = FALSE, allow_null = TRUE)
-  model <- set_default(model, "databricks-dbrx-instruct")
+  model <- set_default(model, "databricks-claude-3-7-sonnet")
   echo <- check_echo(echo)
   if (!is.null(token)) {
     credentials <- function() list(Authorization = paste("Bearer", token))
@@ -177,12 +176,65 @@ method(as_json, list(ProviderDatabricks, ContentText)) <- function(
   x@text
 }
 
-databricks_workspace <- function() {
-  host <- key_get("DATABRICKS_HOST")
-  if (!is.null(host) && !grepl("^https?://", host)) {
-    host <- paste0("https://", host)
+# See: https://docs.databricks.com/aws/en/machine-learning/foundation-model-apis/api-reference#functionobject
+method(as_json, list(ProviderDatabricks, ToolDef)) <- function(provider, x) {
+  # Note: It seems that Databricks doesn't support the "strict" field, despite
+  # what their documentation says. It *is* supported for structured outputs,
+  # though. I suspect a copy & paste error in their docs.
+  compact(list(
+    type = "function",
+    "function" = compact(list(
+      name = x@name,
+      description = x@description,
+      # Use the same parameter encoding as the OpenAI provider, but only if
+      # there actually are parameters.
+      parameters = if (length(x@arguments@properties) != 0)
+        as_json(provider, x@arguments)
+    ))
+  ))
+}
+
+# https://docs.databricks.com/aws/en/machine-learning/foundation-model-apis/api-reference#toolcall
+method(as_json, list(ProviderDatabricks, ContentToolRequest)) <- function(
+  provider,
+  x
+) {
+  # Databricks seems to require encoding empty arguments as an empty
+  # dictionary, rather than an empty array.
+  json_args <- jsonlite::toJSON(set_names(x@arguments))
+  list(
+    id = x@id,
+    `function` = list(name = x@name, arguments = json_args),
+    type = "function"
+  )
+}
+
+databricks_workspace <- function(error_call = caller_env()) {
+  # Default to the well-known DATABRICKS_HOST environment variable.
+  host <- Sys.getenv("DATABRICKS_HOST")
+  if (nchar(host) != 0) {
+    if (!grepl("^https?://", host)) {
+      host <- paste0("https://", host)
+    }
+    return(host)
   }
-  host
+  # If we have no environment variable, check for a Databricks config file
+  # (likely generated via the Databricks CLI or their VS Code extension).
+  cfg_file <- Sys.getenv("DATABRICKS_CONFIG_FILE", "~/.databrickscfg")
+  if (file.exists(cfg_file)) {
+    profile <- Sys.getenv("DATABRICKS_CONFIG_PROFILE", "DEFAULT")
+    host <- databricks_config_field(cfg_file, profile, "host")
+    if (!is.null(host)) {
+      return(host)
+    }
+  }
+  if (is_testing()) {
+    testthat::skip("no Databricks workspace configured")
+  }
+  cli::cli_abort(
+    "No env var {.code DATABRICKS_HOST} set or valid {.file {cfg_file}} found.",
+    call = error_call
+  )
 }
 
 databricks_user_agent <- function() {
@@ -255,11 +307,11 @@ default_databricks_credentials <- function(workspace = databricks_workspace()) {
   # When on desktop, try using the Databricks CLI for auth.
   cli_path <- Sys.getenv("DATABRICKS_CLI_PATH", "databricks")
   if (!is_hosted_session() && nchar(Sys.which(cli_path)) != 0) {
-    token <- databricks_cli_token(cli_path, host)
+    token <- databricks_cli_token(cli_path, workspace)
     if (!is.null(token)) {
       return(function() {
         # Ensure we get an up-to-date token.
-        token <- databricks_cli_token(cli_path, host)
+        token <- databricks_cli_token(cli_path, workspace)
         list(Authorization = paste("Bearer", token))
       })
     }
@@ -337,18 +389,38 @@ databricks_cli_token <- function(cli_path, host) {
 # host = some-host
 # token = some-token
 workbench_databricks_token <- function(host, cfg_file) {
+  databricks_config_field(cfg_file, "workbench", "token")
+}
+
+databricks_config_field <- function(cfg_file, profile, field) {
   cfg <- readLines(cfg_file)
   # We don't attempt a full parse of the INI syntax supported by Databricks
-  # config files, instead relying on the fact that this particular file will
-  # always contain only one section.
-  if (!any(grepl(host, cfg, fixed = TRUE))) {
-    # The configuration doesn't actually apply to this host.
+  # config files and use the following approach to cheat, instead:
+  #
+  # 1. Find the indices of the profile header and the field. If none appear in
+  #    the file, we're done.
+  # 2. Find the index of the next profile header, if any.
+  # 3. Find the index -- if any -- of the field between these headers.
+  # 4. Extract the field's value.
+  profile <- which(grepl(sprintf("[%s]", profile), cfg, fixed = TRUE))
+  if (length(profile) != 1) {
     return(NULL)
   }
-  line <- grepl("token = ", cfg, fixed = TRUE)
-  token <- gsub("token = ", "", cfg[line])
-  if (nchar(token) == 0) {
+  entry <- which(grepl(field, cfg, fixed = TRUE))
+  entry <- entry[entry > profile]
+  if (!any(entry)) {
     return(NULL)
   }
-  token
+  headers <- which(startsWith(cfg, "["))
+  next_profile <- headers[headers > profile]
+  if (!any(next_profile)) {
+    next_profile <- length(cfg) + 1
+  } else {
+    next_profile <- next_profile[1]
+  }
+  entry <- entry[entry < next_profile]
+  if (length(entry) != 1) {
+    return(NULL)
+  }
+  gsub(sprintf("%s\\s*=\\s*(.*)", field), "\\1", cfg[entry])
 }