+++ R/tools-def.R
@@ -76,6 +76,7 @@ tool <- function(
       .name <- unique_tool_name()
     }
   }
+
   ToolDef(
     fun = .fun,
     name = .name,